requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==21.7
gspread==6.1.2
google-auth==2.35.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0